+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: true
 contact_links:
   - name: Community Support
-    url: https://github.com/kessler24/canadata-clean/discussions
+    url: https://github.com/kessler24/canadataClean/discussions
     about: Please ask and answer questions here.
