+++ .github/workflows/build-deploy.yml
@@ -30,7 +30,7 @@ jobs:
       # Run pytest with coverage
       - name: Run tests with coverage
         run: |
-          pytest tests/ --cov=canadata_clean --cov-report=xml
+          pytest tests/ --cov=canadataClean --cov-report=xml
 
   deploy:
     permissions:
@@ -81,7 +81,7 @@ jobs:
             pip install \
             --index-url https://test.pypi.org/simple/ \
             --extra-index-url https://pypi.org/simple \
-            canadata-clean
+            canadataClean
 
       - name: Publish package distributions to GitHub Releases
         uses: python-semantic-release/upload-to-gh-release@main
