+++ pyproject.toml
@@ -11,7 +11,7 @@ requires = ["hatchling"]
 ################################################################################
 
 [project]
-name = "canadata_clean"
+name = "canadataClean"
 # You can chose to use dynamic versioning with hatch or static where you add it manually.
 version = "0.1.0"
 
@@ -41,11 +41,11 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/kessler24/canadata-clean"
-"Source Code" = "https://github.com/kessler24/canadata-clean"
-"Bug Tracker" = "https://github.com/kessler24/canadata-clean/issues"
-Documentation = "https://github.com/kessler24/canadata-clean/blob/main/README.md"
-Download = "https://pypi.org/project/canadata-clean/#files"
+Homepage = "https://github.com/kessler24/canadataClean"
+"Source Code" = "https://github.com/kessler24/canadataClean"
+"Bug Tracker" = "https://github.com/kessler24/canadataClean/issues"
+Documentation = "https://github.com/kessler24/canadataClean/blob/main/README.md"
+Download = "https://pypi.org/project/canadataClean/#files"
 
 [project.optional-dependencies]
 # The groups below should be in the [development-groups] table
@@ -93,7 +93,7 @@ only-packages = true
 # This tells Hatch to build the package from the src/ directory.
 # Read more about src layouts here: https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-structure.html
 [tool.hatch.build.targets.wheel]
-packages = ["src/canadata_clean"]
+packages = ["src/canadataClean"]
 
 
 
@@ -104,8 +104,8 @@ markers = ["raises"] # Tells pytest to not raise a warning if you use @pytest.ma
 
 [tool.coverage.paths]
 source = [
-    "src/canadata_clean",
-    "*/site-packages/canadata_clean",
+    "src/canadataClean",
+    "*/site-packages/canadataClean",
 ]
 
 [tool.coverage.run]
@@ -155,7 +155,7 @@ features = [
 python = ["3.10", "3.11", "3.12", "3.13"]
 
 [tool.hatch.envs.test.scripts]
-run = "pytest {args:--cov=canadata_clean --cov-report=term-missing --cov-report=xml}"
+run = "pytest {args:--cov=canadataClean --cov-report=term-missing --cov-report=xml}"
 
 #--------------- Build and preview your documentation ---------------#
 
@@ -172,7 +172,7 @@ features = [
 # hatch run docs:serve will serve them 'live' on your computer locally
 [tool.hatch.envs.docs.scripts]
 build = ["sphinx-build {args:-W -b html docs docs/_build}"]
-serve = ["sphinx-autobuild docs --watch src/canadata_clean {args:-b html docs/_build/serve}"]
+serve = ["sphinx-autobuild docs --watch src/canadataClean {args:-b html docs/_build/serve}"]
